+++ src/components/Footer/Footer.tsx
@@ -11,7 +11,6 @@ import React, { useCallback, useEffect, useRef } from 'react';
 import OverlayTrigger from 'react-bootstrap/lib/OverlayTrigger';
 import Tooltip from 'react-bootstrap/lib/Tooltip';
 import { useDispatch, useSelector } from 'react-redux';
-import { IState } from 'reducers/types';
 import { selectLoadingFlagByAction } from 'stores/loading/selectors';
 import * as metadataActions from 'stores/metadata/actions';
 import { METADATA_UPDATE_EXECUTE_REQUEST } from 'stores/metadata/constants';
@@ -19,6 +18,7 @@ import {
   getMetadataCurrentVersion,
   getMetadataNewVersion,
 } from 'stores/metadata/selectors';
+import { IState } from 'stores/state';
 
 const FooterGroup = styled.span`
   & + & {
