+++ src/lib/giantswarmClientPatcher.ts
@@ -1,11 +1,11 @@
 import GiantSwarm from 'giantswarm';
 import Auth from 'lib/auth0';
 import { isJwtExpired } from 'lib/helpers';
-import { IState } from 'reducers/types';
 import { AnyAction, Store } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
 import { AuthorizationTypes } from 'shared/constants';
 import { auth0Login } from 'stores/main/actions';
+import { IState } from 'stores/state';
 
 /**
  * This class patches the client's callApi function to check
