+++ src/stores/nodepool/selectors.ts
@@ -1,4 +1,3 @@
-import { IState } from 'reducers/types';
 import { createDeepEqualSelector } from 'selectors/selectorUtils';
 import { INodePool } from 'shared/types';
 import { selectClusterById } from 'stores/cluster/selectors';
@@ -7,6 +6,7 @@ import {
   getMemoryTotalNodePools,
   getNumberOfNodePoolsNodes,
 } from 'stores/cluster/utils';
+import { IState } from 'stores/state';
 
 export function selectNodePools(state: IState): Record<string, INodePool> {
   return state.entities.nodePools.items;
