+++ src/utils/__tests__/clusterUtils.ts
@@ -1,4 +1,3 @@
-import { IState } from 'reducers/types';
 import { Providers } from 'shared/constants';
 import { INodePool, INodePoolStatus, PropertiesOf } from 'shared/types';
 import {
@@ -19,6 +18,7 @@ import {
   isClusterUpdating,
   v4orV5,
 } from 'stores/cluster/utils';
+import { IState } from 'stores/state';
 import {
   v4AWSClusterResponse,
   v4AWSClusterStatusResponse,
